Fixed the create reminder modal some
typescript. ugh.
L#53 if (time.hour) is not a good check for null because if hour was 0
it would return false and not set an hour

Removed the button for the date picker so it can be in-line with time
Now when you click the field it opens the calendar

Gonna work on loading up the reminders now